cmd/roachtest: remove some duplication in ORM tests
We're running several ORM test suites, and each test contains the
copy-pasted logic for parsing the test output and summarizing it.
Now that logic is extracted into a couple of functions which
removes the duplication.

Release note: None
