+++ pkg/cmd/roachtest/psycopg.go
@@ -16,8 +16,6 @@ import (
 	"context"
 	"fmt"
 	"regexp"
-	"sort"
-	"strings"
 )
 
 var psycopgResultRegex = regexp.MustCompile(`(?P<name>.*) \((?P<class>.*)\) \.\.\. (?P<result>[^ ']*)(?: u?['"](?P<reason>.*)['"])?`)
@@ -119,7 +117,7 @@ func registerPsycopg(r *testRegistry) {
 		// Find all the failed and errored tests.
 
 		var failUnexpectedCount, failExpectedCount, ignoredCount, skipCount, unexpectedSkipCount int
-		var passUnexpectedCount, passExpectedCount, notRunCount int
+		var passUnexpectedCount, passExpectedCount int
 		// Put all the results in a giant map of [testname]result
 		results := make(map[string]string)
 		// Current failures are any tests that reported as failed, regardless of if
@@ -178,78 +176,13 @@ func registerPsycopg(r *testRegistry) {
 			}
 		}
 
-		// Collect all the tests that were not run.
-		for test, issue := range expectedFailureList {
-			if _, ok := runTests[test]; ok {
-				continue
-			}
-			allTests = append(allTests, test)
-			results[test] = fmt.Sprintf("--- FAIL: %s - %s (not run)", test, maybeAddGithubLink(issue))
-			notRunCount++
-		}
-
-		// Log all the test results. We re-order the tests alphabetically here
-		// to ensure that we're doing .
-		sort.Strings(allTests)
-		for _, test := range allTests {
-			result, ok := results[test]
-			if !ok {
-				t.Fatalf("can't find %s in test result list", test)
-			}
-			t.l.Printf("%s\n", result)
-		}
-
-		t.l.Printf("------------------------\n")
-
-		var bResults strings.Builder
-		fmt.Fprintf(&bResults, "Tests run on Cockroach %s\n", version)
-		fmt.Fprintf(&bResults, "Tests run against Pyscopg %s\n", latestTag)
-		fmt.Fprintf(&bResults, "%d Total Tests Run\n",
-			passExpectedCount+passUnexpectedCount+failExpectedCount+failUnexpectedCount,
+		summarizeORMTestsResults(
+			t, "psycopg" /* ormName */, blacklistName, expectedFailureList,
+			version, latestTag, currentFailures, allTests, runTests, results,
+			failUnexpectedCount, failExpectedCount, ignoredCount, skipCount,
+			unexpectedSkipCount, passUnexpectedCount, passExpectedCount,
+			nil, /* allIssueHints */
 		)
-
-		p := func(msg string, count int) {
-			testString := "tests"
-			if count == 1 {
-				testString = "test"
-			}
-			fmt.Fprintf(&bResults, "%d %s %s\n", count, testString, msg)
-		}
-		p("passed", passUnexpectedCount+passExpectedCount)
-		p("failed", failUnexpectedCount+failExpectedCount)
-		p("skipped", skipCount)
-		p("ignored", ignoredCount)
-		p("passed unexpectedly", passUnexpectedCount)
-		p("failed unexpectedly", failUnexpectedCount)
-		p("expected failed but skipped", unexpectedSkipCount)
-		p("expected failed but not run", notRunCount)
-
-		fmt.Fprintf(&bResults, "For a full summary look at the psycopg artifacts. \n")
-		t.l.Printf("%s\n", bResults.String())
-		t.l.Printf("------------------------\n")
-
-		if failUnexpectedCount > 0 || passUnexpectedCount > 0 ||
-			notRunCount > 0 || unexpectedSkipCount > 0 {
-			// Create a new psycopg_blacklist so we can easily update this test.
-			sort.Strings(currentFailures)
-			var b strings.Builder
-			fmt.Fprintf(&b, "Here is new psycopg blacklist that can be used to update the test:\n\n")
-			fmt.Fprintf(&b, "var %s = blacklist{\n", blacklistName)
-			for _, test := range currentFailures {
-				issue := expectedFailureList[test]
-				if len(issue) == 0 {
-					issue = "unknown"
-				}
-				fmt.Fprintf(&b, "  \"%s\": \"%s\",\n", test, issue)
-			}
-			fmt.Fprintf(&b, "}\n\n")
-			t.l.Printf("\n\n%s\n\n", b.String())
-			t.l.Printf("------------------------\n")
-			t.Fatalf("\n%s\nAn updated blacklist (%s) is available in the logs\n",
-				bResults.String(),
-				blacklistName,
-			)
-		}
 	}
 
 	r.Add(testSpec{
